requirements.txt
# Middleware Libras — dependências
pyaudiowpatch==0.2.12.6
numpy==1.26.4
openai-whisper==20231117
torch==2.2.2
PyQt6==6.7.0
requests==2.32.3
